move peerinfo util package to cli package
