+++ cli/cmd/peerinfo/mainnet-net-info.json
