+++ cli/cmd/peerinfo/omega-net-info.json
