+++ cli/cmd/peerinfo/peer-info.go
